Topology epochs and topic validation (#17721)
This PR implements two changes in the KIP, and related logic on client & server side.

We replace topology ID with topology epoch. We only include the basic changes to the schema and basic validation for now (fail if topology is changed without a topology epoch bump) - since topology updating will follow later on.
We add status codes for topic validation, and implement the required logic to validate the internal topics.